backend/requirements.txt
PyJWT
Django
django-allauth
cryptography
requests

django-cors-headers
djangorestframework

asgiref
djangorestframework-simplejwt
dj-rest-auth[with_social]

pytz
sqlparse
psycopg2-binary
python-dotenv